Git backend: Add --text when grepping
git-grep sometimes prints "Binary file foo matches" instead
of "foo", even when "-l" is used.  However, it doesn't do this in
all repositories, which is a bit odd.  I can't reproduce this in
a small test case.

sbrowse treats this string as a filename, and fails to open the
file, meaning that later matches are missed.  Work around this by
passing "--text" so that git-grep outputs a filename.